Fix filtering of licenseConfigurations
With the changes of #86, the filtering of configurations broke and test dependencies got reported even though licenseConfigurations was set to Set("compile"). This is due to the fact that the resolved Ivy deps not properly mirror the configurations the dependency was defined for. By selecting the configurations from the report early on, we restore the filtering.